Rename db model to database
