+++ app/scraper/loot/gog_games.py
@@ -6,9 +6,9 @@ from playwright.async_api import Error, Locator, Page
 
 from app.browser import get_new_page
 from app.common import OfferDuration
+from app.database import Offer
 from app.scraper.loot.gog_base import GogBaseScraper
 from app.scraper.loot.scraper import OfferHandler, RawOffer
-from app.sqlalchemy import Offer
 
 logger = logging.getLogger(__name__)
